Add ES module support to JS build in AWS Lambda client
In the kotlin-aws-lambda-runtime-client build.gradle.kts file, ES module support has been added to the JS build of the project. This will allow the usage of ES modules for Node.js in the AWS Lambda client, which can import and export other modules in the project. This change makes our project more compliant with modern JavaScript development practices.