Add suppression for unstable API and generate Javadoc jar
Added a suppression for the unstable API usage warning in the build file. Additionally, set up a task for generating a Javadoc jar file from the project documentation, and added the created Javadoc jar as an artifact in the Maven publication. This change enables us to use unstable APIs without warnings and provides better project documentation for users by including the JavaDocs in the published artifacts.